chore: removed setToUnauthenticated
