+++ src/frontend/AppContainerWrapper.tsx
@@ -22,8 +22,6 @@ import {
   persistNavigationState,
 } from "./Navigation/navigationStateHelperFunctions";
 import { AppNavigator } from "./Navigation/AppNavigator";
-import { AppState, AppStateStatus } from "react-native";
-import { SecurityContext } from "./context/SecurityContext";
 
 // Turn on logging if in debug mode
 if (__DEV__) debug.enable("*");
@@ -35,29 +33,6 @@ const AppContainerWrapper = () => {
   const [initialNavState, setInitialNavState] = React.useState<
     InitialState | "loading" | undefined
   >("loading");
-
-  const { authValuesSet, setToUnauthenticated } = React.useContext(
-    SecurityContext
-  );
-
-  React.useEffect(() => {
-    const appStateListener = AppState.addEventListener(
-      "change",
-      (nextAppState: AppStateStatus) => {
-        if (authValuesSet.passcodeSet) {
-          if (
-            nextAppState === "active" ||
-            nextAppState === "background" ||
-            nextAppState === "inactive"
-          ) {
-            setToUnauthenticated();
-          }
-        }
-      }
-    );
-
-    return () => appStateListener.remove();
-  }, [authValuesSet.passcodeSet, setToUnauthenticated]);
 
   const navRef = useNavigationContainerRef<AppStackList>();
