+++ src/frontend/context/ObservationsContext.tsx
@@ -85,12 +85,14 @@ export const ObservationsProvider = ({
   children,
 }: React.PropsWithChildren<{}>) => {
   const [state, dispatch] = React.useReducer(reducer, defaultContext[0]);
-  const contextValue: ObservationsContextType = React.useMemo(
-    () => [state, dispatch],
-    [state]
-  );
 
   const authState = useAuthState();
+
+  const contextValue: ObservationsContextType = React.useMemo(() => {
+    const derivedState = { ...state };
+    if (authState === "obscured") derivedState.observations = new Map([]);
+    return [derivedState, dispatch];
+  }, [state, authState]);
 
   // This will load observations on first load and reload them every time the
   // value of state.reload changes (dispatch({type: "reload"}) will do this)
@@ -114,15 +116,6 @@ export const ObservationsProvider = ({
       didCancel = true;
     };
   }, [state.reload]);
-
-  React.useEffect(() => {
-    if (authState === "obscured") {
-      dispatch({ type: "reload_success", value: [] });
-      return;
-    }
-
-    dispatch({ type: "reload" });
-  }, [authState]);
 
   return (
     <ObservationsContext.Provider value={contextValue}>
